Updated ToC, +HEARTBEAT option, increased SETTINGS opts len to 32 bits
Added the SETTINGS option sections to the table of contents. Added
a new heartbeat duration option to specify when a client should
send an INFORM message for a heartbeat. Increased the SETTINGS length
to 32 bits to allow a larger amount of data. Filled out the Heartbeat
section and Security Considerations.